+++ tripy/tests/flat_ir/test_constant_deduplication.py
@@ -15,7 +15,7 @@
 
 import pytest
 from tripy.flat_ir.flat_ir import FlatIR
-from tripy.flat_ir.ops.base import FlatIRFunction
+from tripy.flat_ir.function import FlatIRFunction
 from tripy.flat_ir.ops import ConstantOp
 from tripy.flat_ir.tensor import FlatIRTensor
 from tripy.common.device import device
